+++ src/main/java/com/decagon/OakLandv1be/entities/Wallet.java
@@ -5,6 +5,7 @@ import com.fasterxml.jackson.annotation.JsonIgnore;
 import lombok.*;
 
 import javax.persistence.*;
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Entity
@@ -17,7 +18,7 @@ import java.util.Set;
 @Table(name = "wallet_tbl")
 public class Wallet extends BaseEntity{
 
-    private Double accountBalance;
+    private BigDecimal accountBalance;
 
     @Enumerated(EnumType.STRING)
     private BaseCurrency baseCurrency;
